requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-environ==0.11.2
psycopg2==2.9.9
Pillow==10.1.0